my_blog_site/requirements.txt
asgiref==3.2.10
certifi==2020.6.20
chardet==3.0.4
defusedxml==0.6.0
Django==3.0.8
django-allauth==0.42.0
django-ckeditor==6.0.0
django-js-asset==1.2.2
django-markdownx==3.0.1
gunicorn==20.1.0
idna==2.10
importlib-metadata==3.8.0
Markdown==3.3.4
oauthlib==3.1.0
Pillow==8.1.0
psycopg2==2.7.5
python3-openid==3.2.0
pytz==2020.1
requests==2.24.0
requests-oauthlib==1.3.0
six==1.15.0
sqlparse==0.3.1
typing-extensions==3.7.4.3
urllib3==1.25.10
whitenoise==5.2.0
zipp==3.4.1
